edited code

+++ src/DataTypes.java
@@ -3,12 +3,12 @@ import java.util.List;
 public class DataTypes {
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
         }
-        return Long.valueOf(s);
+        return s;
     }
 
 }
